Show use of more unique trace number to prevent errors
When we use time() as the traceNumber and more than one transaction is processed per second, the first transaction processes normally, then the second (using the same traceNumber)  is sent. Orbital returns the result of the first transaction (because it matches the traceNumber) rather than actually processing the second one.

Switch to using hrtime(true) so we get a more unique traceNumber almost to the microsecond.